Support '/' in --output-template
This allows to specify subdirectories with the output template definition using
'/'.  The logic of title formatter is modified so that the '/'s from metadata
is replaced with '_' while the user specified '/'s are preserved. Nested
subdirectories are also supported.

For example: `--output-template '${series}/${title} --create-dirs` will store
files as '${title} to subdirectories defined by the `${series}` metadata.