yledl\downloader.py	 comparison-of-constants
The alert actually found a bug, which was verified with aajanki
limits.start_position might be None so (limits.start_position or 0) gives it a default of 0 in this case.
The added parenthesis force evaluation order to implement the desired order and avoids the unintentional comparison of constants.
The external added parenthesis in ((limits.start_position or 0) > 0)  is for clarity