Improve FFmpeg version checker regular expression
When FFmpeg is built from Git sources, the version is prefixed with `n`:

    $ ffmpeg -loglevel quiet -version
    ffmpeg version n7.1.1 Copyright (c) 2000-2025 the FFmpeg developers

When FFmpeg is built from release tar balls, there is no prefix:

    $ ffmpeg -loglevel quiet -version
    ffmpeg version 7.1.1 Copyright (c) 2000-2025 the FFmpeg developers

This commit ensures that both cases are handled correctly.

The practical implication was that subtitles failed to download because
the `n` prefix broke the test introduced in a previous commit[1].

Note that this bug affected only FFmpeg packages that were built
directly from Git sources. One example is the FFmpeg in Arch Linux
package repositories[2].

References:
[1] dce1110fcb7a1304fd3843f8ac3a613d95a181d9
[2] https://gitlab.archlinux.org/archlinux/packaging/packages/ffmpeg/-/blob/1e27244932c2c61ab340b357ff4ac8dd0065e3af/PKGBUILD#L123