Make tests compatible with Python 3.15
Python 3.15 changes `strptime` behavior such that empty string is
allowed for `%z` [1]. However, in `test_timestamp.py` there is
an assertion that explicitly tests that a timestamp that is otherwise
valid but uses an empty string for `%z` is rejected.

Make the test suite pass for both Python 3.15 and older Pythons by
simply removing that assertion. As the test cases for success are not
modified, specification for valid input is not modified. Just one case
that turns out to be valid in new Python versions is removed from the
set of explictly rejected cases.

This issue case up when Fedora test build for Python 3.15 failed due to
this issue [2].

[1]: https://github.com/python/cpython/issues/122781
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=2424323